dataHub Class

dataHub class, constructor, and methods.

Usage

dataHub(BFC)

dataHub(BFC)

# S4 method for dataHub
show(object)

dataNames(object)

dataParams(object)

dataNotes(object)

dataPaths(object)

dataYml(object)

dataTags(object)

# S4 method for dataHub
dataTags(object)

dataTags(object, append = TRUE) <- value

# S4 method for dataHub
dataTags(object, append = FALSE) <- value

# S4 method for dataHub,ANY,ANY,ANY
[(x, i, j, drop)

# S4 method for dataHub,ANY,ANY,ANY
[(x, i, j) <- value

# S4 method for dataHub
c(x, ...)

toList(
  x,
  listNames = NULL,
  format = c("list", "json", "yaml"),
  type = NULL,
  file = character()
)

Arguments

BFC

A BiocFileCache object created for data and recipes.

object

A dataHub object.

append

Whether to append new tag or replace all tags.

value

A dataHub object

x

A dataHub object.

i

The integer index of the dataHub object, or a logical vector same length as the dataHub object.

j

inherited from [ generic.

drop

Inherited from [ generic.

...

More dataHub objects to combine.

listNames

A vector of names for the output list.

format

can be "list", "json" or "yaml". Supports partial match. Default is list.

type

The type of workflow input list, such as cwl.

file

The file name to save the data list in required format. The data extension needs to be included, e.g., ".json" or ".yml".

Value

dataHub: a dataHub object.

dataNames: the names of datasets in dataHub object.

dataParams: the data recipe parameter values for datasets in dataHub object.

dataNotes: the notes of datasets in dataHub object.

dataPaths: the file paths of datasets in dataHub object.

dataYml: the yaml file paths of datasets in dataHub object.

dataTags: the tags of datasets in dataHub object.

toList: A list of datasets in specific format, and a file if file argument is specified.

Examples

outdir <- file.path(tempdir(), "SharedData")
dataUpdate(outdir, cloud = TRUE)
#> 
#> Updating data record...
#> 35c86a4f3633_GRCh38.primary_assembly.genome.fa.1.bt2 added
#> 35c86144ba7c7_GRCh38.primary_assembly.genome.fa.2.bt2 added
#> 35c8611076c9_GRCh38.primary_assembly.genome.fa.3.bt2 added
#> 35c8673a08e08_GRCh38.primary_assembly.genome.fa.4.bt2 added
#> 35c8643415668_GRCh38.primary_assembly.genome.fa.rev.1.bt2 added
#> 35c862690a05a_GRCh38.primary_assembly.genome.fa.rev.2.bt2 added
#> 35c8664d3620b_outfile.txt added
#> 35c8656583273_GRCh37_to_GRCh38.chain added
#> 35c861497364b_GRCh37_to_NCBI34.chain added
#> 35c866fbab21b_GRCh37_to_NCBI35.chain added
#> 35c866bb38cc0_GRCh37_to_NCBI36.chain added
#> 35c866272c88a_GRCh38_to_GRCh37.chain added
#> 35c864a066867_GRCh38_to_NCBI34.chain added
#> 35c86201c403a_GRCh38_to_NCBI35.chain added
#> 35c862d2ca0a0_GRCh38_to_NCBI36.chain added
#> 35c866c1ecab0_NCBI34_to_GRCh37.chain added
#> 35c861639db7d_NCBI34_to_GRCh38.chain added
#> 35c864a9afee9_NCBI35_to_GRCh37.chain added
#> 35c86ab4fbf0_NCBI35_to_GRCh38.chain added
#> 35c8676b9d409_NCBI36_to_GRCh37.chain added
#> 35c866e15c74b_NCBI36_to_GRCh38.chain added
#> 35c86273b4a7d_GRCm38_to_NCBIM36.chain added
#> 35c865092c6e_GRCm38_to_NCBIM37.chain added
#> 35c86e89b952_NCBIM36_to_GRCm38.chain added
#> 35c867756569c_NCBIM37_to_GRCm38.chain added
#> 35c8631187d53_1000G_omni2.5.b37.vcf.gz added
#> 35c863e2ce24e_1000G_omni2.5.b37.vcf.gz.tbi added
#> 35c86b3b310c_Mills_and_1000G_gold_standard.indels.b37.vcf.gz added
#> 35c8661af5314_Mills_and_1000G_gold_standard.indels.b37.vcf.gz.tbi added
#> 35c862f733cd2_1000G_omni2.5.hg38.vcf.gz added
#> 35c864ff39ec3_1000G_omni2.5.hg38.vcf.gz.tbi added
#> 35c866bfe8948_Mills_and_1000G_gold_standard.indels.hg38.vcf.gz added
#> 35c8643bee499_Mills_and_1000G_gold_standard.indels.hg38.vcf.gz.tbi added
#> 35c865104158c_af-only-gnomad.raw.sites.vcf added
#> 35c865f9f1750_af-only-gnomad.raw.sites.vcf.idx added
#> 35c867003b02_Mutect2-exome-panel.vcf.idx added
#> 35c867794b5e6_Mutect2-WGS-panel-b37.vcf added
#> 35c864472795b_Mutect2-WGS-panel-b37.vcf.idx added
#> 35c865d586d75_small_exac_common_3.vcf added
#> 35c86c2bec31_small_exac_common_3.vcf.idx added
#> 35c86342d2b77_1000g_pon.hg38.vcf.gz added
#> 35c86490bfa35_1000g_pon.hg38.vcf.gz.tbi added
#> 35c866e9eb4bb_af-only-gnomad.hg38.vcf.gz added
#> 35c867e3393de_af-only-gnomad.hg38.vcf.gz.tbi added
#> 35c8669283a6f_small_exac_common_3.hg38.vcf.gz added
#> 35c861bcb555b_small_exac_common_3.hg38.vcf.gz.tbi added
#> 35c866a525e8f_gencode.v41.annotation.gtf added
#> 35c867f6215ec_gencode.v42.annotation.gtf added
#> 35c8666665444_gencode.vM30.annotation.gtf added
#> 35c8675075a7f_gencode.vM31.annotation.gtf added
#> 35c86761be9f6_gencode.v41.transcripts.fa added
#> 35c86547c1b90_gencode.v41.transcripts.fa.fai added
#> 35c861c42a4fc_gencode.v42.transcripts.fa added
#> 35c867b251664_gencode.v42.transcripts.fa.fai added
#> 35c866305d4e2_gencode.vM30.pc_transcripts.fa added
#> 35c861398fb98_gencode.vM30.pc_transcripts.fa.fai added
#> 35c862c3d93b8_gencode.vM31.pc_transcripts.fa added
#> 35c862132b730_gencode.vM31.pc_transcripts.fa.fai added
#> 35c861ed42ca4_GRCh38.primary_assembly.genome.fa.1.ht2 added
#> 35c86dece6cc_GRCh38.primary_assembly.genome.fa.2.ht2 added
#> 35c8650a5f402_GRCh38.primary_assembly.genome.fa.3.ht2 added
#> 35c866ec7cb67_GRCh38.primary_assembly.genome.fa.4.ht2 added
#> 35c8679eb7014_GRCh38.primary_assembly.genome.fa.5.ht2 added
#> 35c861464d89b_GRCh38.primary_assembly.genome.fa.6.ht2 added
#> 35c863fcbe0f3_GRCh38.primary_assembly.genome.fa.7.ht2 added
#> 35c86598a8764_GRCh38.primary_assembly.genome.fa.8.ht2 added
#> 35c861b65139d_GRCh38_full_analysis_set_plus_decoy_hla.fa.fai added
#> 35c86376096d9_GRCh38_full_analysis_set_plus_decoy_hla.fa.amb added
#> 35c861dfd00c0_GRCh38_full_analysis_set_plus_decoy_hla.fa.ann added
#> 35c8678bd8113_GRCh38_full_analysis_set_plus_decoy_hla.fa.bwt added
#> 35c86438c830a_GRCh38_full_analysis_set_plus_decoy_hla.fa.pac added
#> 35c86522a2c37_GRCh38_full_analysis_set_plus_decoy_hla.fa.sa added
#> 35c8641c97b48_GRCh38_full_analysis_set_plus_decoy_hla.fa added
#> 35c86322b37c5_GRCh38.primary_assembly.genome.fa.fai added
#> 35c86505dc015_GRCh38.primary_assembly.genome.fa.amb added
#> 35c862af1b5b8_GRCh38.primary_assembly.genome.fa.ann added
#> 35c864df68d20_GRCh38.primary_assembly.genome.fa.bwt added
#> 35c863ab01ea4_GRCh38.primary_assembly.genome.fa.pac added
#> 35c862a53cba4_GRCh38.primary_assembly.genome.fa.sa added
#> 35c86345ce164_GRCh38.primary_assembly.genome.fa added
#> 35c862fb77923_hs37d5.fa.fai added
#> 35c86206fb59a_hs37d5.fa.amb added
#> 35c868d8fcf4_hs37d5.fa.ann added
#> 35c864bfa1e1f_hs37d5.fa.bwt added
#> 35c861b94cbff_hs37d5.fa.pac added
#> 35c866bded1d6_hs37d5.fa.sa added
#> 35c865f9319b7_hs37d5.fa added
#> 35c8647d25fb7_complete_ref_lens.bin added
#> 35c86d118906_ctable.bin added
#> 35c867e67465b_ctg_offsets.bin added
#> 35c8655bf4683_duplicate_clusters.tsv added
#> 35c865db77d08_info.json added
#> 35c866d2f11c2_mphf.bin added
#> 35c864faab698_pos.bin added
#> 35c86721c55a4_pre_indexing.log added
#> 35c862cfaf2b5_rank.bin added
#> 35c8629353dfc_ref_indexing.log added
#> 35c86d816941_refAccumLengths.bin added
#> 35c86645b898e_reflengths.bin added
#> 35c8647323ebc_refseq.bin added
#> 35c8663eea54_seq.bin added
#> 35c8627e80c98_versionInfo.json added
#> 35c86195c6af3_salmon_index added
#> 35c864808659d_chrLength.txt added
#> 35c865a13445d_chrName.txt added
#> 35c8669ba2b09_chrNameLength.txt added
#> 35c8672fa1b55_chrStart.txt added
#> 35c862809d17d_exonGeTrInfo.tab added
#> 35c86246a49ad_exonInfo.tab added
#> 35c861d4de6f9_geneInfo.tab added
#> 35c865c66b2e2_Genome added
#> 35c865421c2d1_genomeParameters.txt added
#> 35c863dbd9c94_Log.out added
#> 35c86653fafd6_SA added
#> 35c86201be0f0_SAindex added
#> 35c8659526893_sjdbInfo.txt added
#> 35c86511e81ad_sjdbList.fromGTF.out.tab added
#> 35c867faefaa8_sjdbList.out.tab added
#> 35c862124c84a_transcriptInfo.tab added
#> 35c865e300ab3_GRCh38.GENCODE.v42_100 added
#> 35c867e164103_knownGene_hg38.sql added
#> 35c8676e40ecd_knownGene_hg38.txt added
#> 35c863be787bc_refGene_hg38.sql added
#> 35c866b4552c6_refGene_hg38.txt added
#> 35c86468ec565_knownGene_mm39.sql added
#> 35c862e03dd60_knownGene_mm39.txt added
#> 35c861840457b_refGene_mm39.sql added
#> 35c866fc40362_refGene_mm39.txt added
#> dataHub with 144 records
#> cache path:  /home/qhu/.cache/R/ReUseData 
#> # dataUpdate() to update the local data cache
#> # dataSearch() to query a specific dataset
#> # Additional information can be retrieved using: 
#> # dataNames(), dataParams(), dataNotes(), dataPaths(), dataTag() or mcols()
#> 
#>             name                   
#>   BFC899  | GRCh37_to_GRCh38.chain 
#>   BFC1927 | GRCm38_to_NCBIM36.chain
#>   BFC2825 | GRCh37_to_GRCh38.chain 
#>   BFC2826 | outfile.txt            
#>   BFC2955 | outfile.txt            
#>   ...       ...                    
#>   BFC4626 | refGene_hg38.txt       
#>   BFC4627 | knownGene_mm39.sql     
#>   BFC4628 | knownGene_mm39.txt     
#>   BFC4629 | refGene_mm39.sql       
#>   BFC4630 | refGene_mm39.txt       
#>           Path                                                               
#>   BFC899  /tmp/RtmpMZs31N/gcpData/GRCh37_to_GRCh38.chain                     
#>   BFC1927 /tmp/Rtmp4dXPOh/gcpData/GRCm38_to_NCBIM36.chain                    
#>   BFC2825 /tmp/Rtmp4dXPOh/gcpData/GRCh37_to_GRCh38.chain                     
#>   BFC2826 /tmp/RtmpkdDQm8/working_dir/RtmpjAGUWK/test_SharedData/outfile.txt 
#>   BFC2955 /tmp/RtmpRwEJgD/working_dir/Rtmpilpixw/test_SharedData/outfile.txt 
#>   ...     ...                                                                
#>   BFC4626 https://storage.googleapis.com/reusedata/ucsc_database/refGene_h...
#>   BFC4627 https://storage.googleapis.com/reusedata/ucsc_database/knownGene...
#>   BFC4628 https://storage.googleapis.com/reusedata/ucsc_database/knownGene...
#>   BFC4629 https://storage.googleapis.com/reusedata/ucsc_database/refGene_m...
#>   BFC4630 https://storage.googleapis.com/reusedata/ucsc_database/refGene_m...
dd <- dataSearch(c("liftover", "GRCh38"))
dataNames(dd)
#>  [1] "GRCh37_to_GRCh38.chain" "GRCh37_to_GRCh38.chain" "GRCh37_to_GRCh38.chain"
#>  [4] "GRCh37_to_GRCh38.chain" "GRCh37_to_GRCh38.chain" "GRCh37_to_GRCh38.chain"
#>  [7] "GRCh37_to_GRCh38.chain" "GRCh37_to_GRCh38.chain" "GRCh37_to_GRCh38.chain"
#> [10] "GRCh37_to_GRCh38.chain" "GRCh37_to_GRCh38.chain" "GRCh37_to_GRCh38.chain"
#> [13] "GRCh37_to_GRCh38.chain" "GRCh38_to_GRCh37.chain" "GRCh38_to_NCBI34.chain"
#> [16] "GRCh38_to_NCBI35.chain" "GRCh38_to_NCBI36.chain" "NCBI34_to_GRCh38.chain"
#> [19] "NCBI35_to_GRCh38.chain" "NCBI36_to_GRCh38.chain"
dataParams(dd)
#>  [1] "species: human; from: GRCh37; to: GRCh38"
#>  [2] "species: human; from: GRCh37; to: GRCh38"
#>  [3] "species: human; from: GRCh37; to: GRCh38"
#>  [4] "species: human; from: GRCh37; to: GRCh38"
#>  [5] "species: human; from: GRCh37; to: GRCh38"
#>  [6] "species: human; from: GRCh37; to: GRCh38"
#>  [7] "species: human; from: GRCh37; to: GRCh38"
#>  [8] "species: human; from: GRCh37; to: GRCh38"
#>  [9] "species: human; from: GRCh37; to: GRCh38"
#> [10] "species: human; from: GRCh37; to: GRCh38"
#> [11] "species: human; from: GRCh37; to: GRCh38"
#> [12] "species: human; from: GRCh37; to: GRCh38"
#> [13] "species: human; from: GRCh37; to: GRCh38"
#> [14] "species: human; from: GRCh38; to: GRCh37"
#> [15] "species: human; from: GRCh38; to: NCBI34"
#> [16] "species: human; from: GRCh38; to: NCBI35"
#> [17] "species: human; from: GRCh38; to: NCBI36"
#> [18] "species: human; from: NCBI34; to: GRCh38"
#> [19] "species: human; from: NCBI35; to: GRCh38"
#> [20] "species: human; from: NCBI36; to: GRCh38"
dataNotes(dd)
#>  [1] "ensembl liftover human GRCh37 GRCh38"
#>  [2] "ensembl liftover human GRCh37 GRCh38"
#>  [3] "ensembl liftover human GRCh37 GRCh38"
#>  [4] "ensembl liftover human GRCh37 GRCh38"
#>  [5] "ensembl liftover human GRCh37 GRCh38"
#>  [6] "ensembl liftover human GRCh37 GRCh38"
#>  [7] "ensembl liftover human GRCh37 GRCh38"
#>  [8] "ensembl liftover human GRCh37 GRCh38"
#>  [9] "ensembl liftover human GRCh37 GRCh38"
#> [10] "ensembl liftover human GRCh37 GRCh38"
#> [11] "ensembl liftover human GRCh37 GRCh38"
#> [12] "ensembl liftover human GRCh37 GRCh38"
#> [13] "ensembl liftover human GRCh37 GRCh38"
#> [14] "ensembl liftover human GRCh38 GRCh37"
#> [15] "ensembl liftover human GRCh38 NCBI34"
#> [16] "ensembl liftover human GRCh38 NCBI35"
#> [17] "ensembl liftover human GRCh38 NCBI36"
#> [18] "ensembl liftover human NCBI34 GRCh38"
#> [19] "ensembl liftover human NCBI35 GRCh38"
#> [20] "ensembl liftover human NCBI36 GRCh38"
dataTags(dd)
#>  [1] "" "" "" "" "" "" "" "" "" "" "" "" "" "" "" "" "" "" "" ""
dataYml(dd)
#>  [1] "/tmp/RtmpMZs31N/gcpData/genome_liftover_human_GRCh37_GRCh38.yml"                                  
#>  [2] "/tmp/Rtmp4dXPOh/gcpData/genome_liftover_human_GRCh37_GRCh38.yml"                                  
#>  [3] "/tmp/RtmpMF39J6/gcpData/genome_liftover_human_GRCh37_GRCh38.yml"                                  
#>  [4] "/tmp/Rtmp8mVqnL/gcpData/genome_liftover_human_GRCh37_GRCh38.yml"                                  
#>  [5] "/tmp/RtmpKEpli0/gcpData/genome_liftover_human_GRCh37_GRCh38.yml"                                  
#>  [6] "/tmp/Rtmpm1thzu/gcpData/genome_liftover_human_GRCh37_GRCh38.yml"                                  
#>  [7] "/tmp/Rtmp2f3PBx/gcpData/genome_liftover_human_GRCh37_GRCh38.yml"                                  
#>  [8] "/tmp/RtmpjrG1vn/gcpData/genome_liftover_human_GRCh37_GRCh38.yml"                                  
#>  [9] "/tmp/RtmpldLvSD/gcpData/genome_liftover_human_GRCh37_GRCh38.yml"                                  
#> [10] "/tmp/RtmpTvoXjc/gcpData/genome_liftover_human_GRCh37_GRCh38.yml"                                  
#> [11] "/tmp/Rtmp852CtF/gcpData/genome_liftover_human_GRCh37_GRCh38.yml"                                  
#> [12] "/tmp/Rtmpx6qi6B/gcpData/genome_liftover_human_GRCh37_GRCh38.yml"                                  
#> [13] "https://storage.googleapis.com/reusedata/ensembl_liftover/genome_liftover_human_GRCh37_GRCh38.yml"
#> [14] "https://storage.googleapis.com/reusedata/ensembl_liftover/genome_liftover_human_GRCh38_GRCh37.yml"
#> [15] "https://storage.googleapis.com/reusedata/ensembl_liftover/genome_liftover_human_GRCh38_NCBI34.yml"
#> [16] "https://storage.googleapis.com/reusedata/ensembl_liftover/genome_liftover_human_GRCh38_NCBI35.yml"
#> [17] "https://storage.googleapis.com/reusedata/ensembl_liftover/genome_liftover_human_GRCh38_NCBI36.yml"
#> [18] "https://storage.googleapis.com/reusedata/ensembl_liftover/genome_liftover_human_NCBI34_GRCh38.yml"
#> [19] "https://storage.googleapis.com/reusedata/ensembl_liftover/genome_liftover_human_NCBI35_GRCh38.yml"
#> [20] "https://storage.googleapis.com/reusedata/ensembl_liftover/genome_liftover_human_NCBI36_GRCh38.yml"
toList(dd)
#> $GRCh37_to_GRCh38.chain
#> [1] "/tmp/RtmpMZs31N/gcpData/GRCh37_to_GRCh38.chain"
#> 
#> $GRCh37_to_GRCh38.chain
#> [1] "/tmp/Rtmp4dXPOh/gcpData/GRCh37_to_GRCh38.chain"
#> 
#> $GRCh37_to_GRCh38.chain
#> [1] "/tmp/RtmpMF39J6/gcpData/GRCh37_to_GRCh38.chain"
#> 
#> $GRCh37_to_GRCh38.chain
#> [1] "/tmp/Rtmp8mVqnL/gcpData/GRCh37_to_GRCh38.chain"
#> 
#> $GRCh37_to_GRCh38.chain
#> [1] "/tmp/RtmpKEpli0/gcpData/GRCh37_to_GRCh38.chain"
#> 
#> $GRCh37_to_GRCh38.chain
#> [1] "/tmp/Rtmpm1thzu/gcpData/GRCh37_to_GRCh38.chain"
#> 
#> $GRCh37_to_GRCh38.chain
#> [1] "/tmp/Rtmp2f3PBx/gcpData/GRCh37_to_GRCh38.chain"
#> 
#> $GRCh37_to_GRCh38.chain
#> [1] "/tmp/RtmpjrG1vn/gcpData/GRCh37_to_GRCh38.chain"
#> 
#> $GRCh37_to_GRCh38.chain
#> [1] "/tmp/RtmpldLvSD/gcpData/GRCh37_to_GRCh38.chain"
#> 
#> $GRCh37_to_GRCh38.chain
#> [1] "/tmp/RtmpTvoXjc/gcpData/GRCh37_to_GRCh38.chain"
#> 
#> $GRCh37_to_GRCh38.chain
#> [1] "/tmp/Rtmp852CtF/gcpData/GRCh37_to_GRCh38.chain"
#> 
#> $GRCh37_to_GRCh38.chain
#> [1] "/tmp/Rtmpx6qi6B/gcpData/GRCh37_to_GRCh38.chain"
#> 
#> $GRCh37_to_GRCh38.chain
#> [1] "gs://reusedata/ensembl_liftover/GRCh37_to_GRCh38.chain"
#> 
#> $GRCh38_to_GRCh37.chain
#> [1] "gs://reusedata/ensembl_liftover/GRCh38_to_GRCh37.chain"
#> 
#> $GRCh38_to_NCBI34.chain
#> [1] "gs://reusedata/ensembl_liftover/GRCh38_to_NCBI34.chain"
#> 
#> $GRCh38_to_NCBI35.chain
#> [1] "gs://reusedata/ensembl_liftover/GRCh38_to_NCBI35.chain"
#> 
#> $GRCh38_to_NCBI36.chain
#> [1] "gs://reusedata/ensembl_liftover/GRCh38_to_NCBI36.chain"
#> 
#> $NCBI34_to_GRCh38.chain
#> [1] "gs://reusedata/ensembl_liftover/NCBI34_to_GRCh38.chain"
#> 
#> $NCBI35_to_GRCh38.chain
#> [1] "gs://reusedata/ensembl_liftover/NCBI35_to_GRCh38.chain"
#> 
#> $NCBI36_to_GRCh38.chain
#> [1] "gs://reusedata/ensembl_liftover/NCBI36_to_GRCh38.chain"
#> 
toList(dd, format = "yaml")
#> [1] "GRCh37_to_GRCh38.chain: /tmp/RtmpMZs31N/gcpData/GRCh37_to_GRCh38.chain\nGRCh37_to_GRCh38.chain: /tmp/Rtmp4dXPOh/gcpData/GRCh37_to_GRCh38.chain\nGRCh37_to_GRCh38.chain: /tmp/RtmpMF39J6/gcpData/GRCh37_to_GRCh38.chain\nGRCh37_to_GRCh38.chain: /tmp/Rtmp8mVqnL/gcpData/GRCh37_to_GRCh38.chain\nGRCh37_to_GRCh38.chain: /tmp/RtmpKEpli0/gcpData/GRCh37_to_GRCh38.chain\nGRCh37_to_GRCh38.chain: /tmp/Rtmpm1thzu/gcpData/GRCh37_to_GRCh38.chain\nGRCh37_to_GRCh38.chain: /tmp/Rtmp2f3PBx/gcpData/GRCh37_to_GRCh38.chain\nGRCh37_to_GRCh38.chain: /tmp/RtmpjrG1vn/gcpData/GRCh37_to_GRCh38.chain\nGRCh37_to_GRCh38.chain: /tmp/RtmpldLvSD/gcpData/GRCh37_to_GRCh38.chain\nGRCh37_to_GRCh38.chain: /tmp/RtmpTvoXjc/gcpData/GRCh37_to_GRCh38.chain\nGRCh37_to_GRCh38.chain: /tmp/Rtmp852CtF/gcpData/GRCh37_to_GRCh38.chain\nGRCh37_to_GRCh38.chain: /tmp/Rtmpx6qi6B/gcpData/GRCh37_to_GRCh38.chain\nGRCh37_to_GRCh38.chain: gs://reusedata/ensembl_liftover/GRCh37_to_GRCh38.chain\nGRCh38_to_GRCh37.chain: gs://reusedata/ensembl_liftover/GRCh38_to_GRCh37.chain\nGRCh38_to_NCBI34.chain: gs://reusedata/ensembl_liftover/GRCh38_to_NCBI34.chain\nGRCh38_to_NCBI35.chain: gs://reusedata/ensembl_liftover/GRCh38_to_NCBI35.chain\nGRCh38_to_NCBI36.chain: gs://reusedata/ensembl_liftover/GRCh38_to_NCBI36.chain\nNCBI34_to_GRCh38.chain: gs://reusedata/ensembl_liftover/NCBI34_to_GRCh38.chain\nNCBI35_to_GRCh38.chain: gs://reusedata/ensembl_liftover/NCBI35_to_GRCh38.chain\nNCBI36_to_GRCh38.chain: gs://reusedata/ensembl_liftover/NCBI36_to_GRCh38.chain"
toList(dd, format = "json", file = tempfile())
#> File is saved as: "/tmp/Rtmp7Lq2kA/file35c863b8546a1"
#> {
#>   "GRCh37_to_GRCh38.chain": "/tmp/RtmpMZs31N/gcpData/GRCh37_to_GRCh38.chain",
#>   "GRCh37_to_GRCh38.chain.1": "/tmp/Rtmp4dXPOh/gcpData/GRCh37_to_GRCh38.chain",
#>   "GRCh37_to_GRCh38.chain.2": "/tmp/RtmpMF39J6/gcpData/GRCh37_to_GRCh38.chain",
#>   "GRCh37_to_GRCh38.chain.3": "/tmp/Rtmp8mVqnL/gcpData/GRCh37_to_GRCh38.chain",
#>   "GRCh37_to_GRCh38.chain.4": "/tmp/RtmpKEpli0/gcpData/GRCh37_to_GRCh38.chain",
#>   "GRCh37_to_GRCh38.chain.5": "/tmp/Rtmpm1thzu/gcpData/GRCh37_to_GRCh38.chain",
#>   "GRCh37_to_GRCh38.chain.6": "/tmp/Rtmp2f3PBx/gcpData/GRCh37_to_GRCh38.chain",
#>   "GRCh37_to_GRCh38.chain.7": "/tmp/RtmpjrG1vn/gcpData/GRCh37_to_GRCh38.chain",
#>   "GRCh37_to_GRCh38.chain.8": "/tmp/RtmpldLvSD/gcpData/GRCh37_to_GRCh38.chain",
#>   "GRCh37_to_GRCh38.chain.9": "/tmp/RtmpTvoXjc/gcpData/GRCh37_to_GRCh38.chain",
#>   "GRCh37_to_GRCh38.chain.10": "/tmp/Rtmp852CtF/gcpData/GRCh37_to_GRCh38.chain",
#>   "GRCh37_to_GRCh38.chain.11": "/tmp/Rtmpx6qi6B/gcpData/GRCh37_to_GRCh38.chain",
#>   "GRCh37_to_GRCh38.chain.12": "gs://reusedata/ensembl_liftover/GRCh37_to_GRCh38.chain",
#>   "GRCh38_to_GRCh37.chain": "gs://reusedata/ensembl_liftover/GRCh38_to_GRCh37.chain",
#>   "GRCh38_to_NCBI34.chain": "gs://reusedata/ensembl_liftover/GRCh38_to_NCBI34.chain",
#>   "GRCh38_to_NCBI35.chain": "gs://reusedata/ensembl_liftover/GRCh38_to_NCBI35.chain",
#>   "GRCh38_to_NCBI36.chain": "gs://reusedata/ensembl_liftover/GRCh38_to_NCBI36.chain",
#>   "NCBI34_to_GRCh38.chain": "gs://reusedata/ensembl_liftover/NCBI34_to_GRCh38.chain",
#>   "NCBI35_to_GRCh38.chain": "gs://reusedata/ensembl_liftover/NCBI35_to_GRCh38.chain",
#>   "NCBI36_to_GRCh38.chain": "gs://reusedata/ensembl_liftover/NCBI36_to_GRCh38.chain"
#> }