dataSearch search data in local data caching system

dataSearch search data in local data caching system

Usage

dataSearch(keywords = character(), cachePath = "ReUseData")

Arguments

keywords

character vector of keywords to be matched to the local datasets. It matches the "notes" when generating the data using getData(notes = ). Keywords can be a tag with the data in #tag format. If not specified, function returns the full data list.

cachePath

A character string for the data cache. Must match the one specified in dataUpdate(). Default is "ReUseData".

Value

a dataHub object containing the information about local data cache, e.g., data name, data path, etc.

Examples

dataSearch()
#> dataHub with 144 records
#> cache path:  /home/qhu/.cache/R/ReUseData 
#> # dataUpdate() to update the local data cache
#> # dataSearch() to query a specific dataset
#> # Additional information can be retrieved using: 
#> # dataNames(), dataParams(), dataNotes(), dataPaths(), dataTag() or mcols()
#> 
#>             name                   
#>   BFC899  | GRCh37_to_GRCh38.chain 
#>   BFC1927 | GRCm38_to_NCBIM36.chain
#>   BFC2825 | GRCh37_to_GRCh38.chain 
#>   BFC2826 | outfile.txt            
#>   BFC2955 | outfile.txt            
#>   ...       ...                    
#>   BFC4626 | refGene_hg38.txt       
#>   BFC4627 | knownGene_mm39.sql     
#>   BFC4628 | knownGene_mm39.txt     
#>   BFC4629 | refGene_mm39.sql       
#>   BFC4630 | refGene_mm39.txt       
#>           Path                                                               
#>   BFC899  /tmp/RtmpMZs31N/gcpData/GRCh37_to_GRCh38.chain                     
#>   BFC1927 /tmp/Rtmp4dXPOh/gcpData/GRCm38_to_NCBIM36.chain                    
#>   BFC2825 /tmp/Rtmp4dXPOh/gcpData/GRCh37_to_GRCh38.chain                     
#>   BFC2826 /tmp/RtmpkdDQm8/working_dir/RtmpjAGUWK/test_SharedData/outfile.txt 
#>   BFC2955 /tmp/RtmpRwEJgD/working_dir/Rtmpilpixw/test_SharedData/outfile.txt 
#>   ...     ...                                                                
#>   BFC4626 https://storage.googleapis.com/reusedata/ucsc_database/refGene_h...
#>   BFC4627 https://storage.googleapis.com/reusedata/ucsc_database/knownGene...
#>   BFC4628 https://storage.googleapis.com/reusedata/ucsc_database/knownGene...
#>   BFC4629 https://storage.googleapis.com/reusedata/ucsc_database/refGene_m...
#>   BFC4630 https://storage.googleapis.com/reusedata/ucsc_database/refGene_m...
dataSearch(c("gencode")) 
#> dataHub with 52 records
#> cache path:  /home/qhu/.cache/R/ReUseData 
#> # dataUpdate() to update the local data cache
#> # dataSearch() to query a specific dataset
#> # Additional information can be retrieved using: 
#> # dataNames(), dataParams(), dataNotes(), dataPaths(), dataTag() or mcols()
#> 
#>             name                       
#>   BFC4549 | gencode.v41.annotation.gtf 
#>   BFC4550 | gencode.v42.annotation.gtf 
#>   BFC4551 | gencode.vM30.annotation.gtf
#>   BFC4552 | gencode.vM31.annotation.gtf
#>   BFC4553 | gencode.v41.transcripts.fa 
#>   ...       ...                        
#>   BFC4618 | sjdbInfo.txt               
#>   BFC4619 | sjdbList.fromGTF.out.tab   
#>   BFC4620 | sjdbList.out.tab           
#>   BFC4621 | transcriptInfo.tab         
#>   BFC4622 | GRCh38.GENCODE.v42_100     
#>           Path                                                               
#>   BFC4549 https://storage.googleapis.com/reusedata/gencode_annotation/genc...
#>   BFC4550 https://storage.googleapis.com/reusedata/gencode_annotation/genc...
#>   BFC4551 https://storage.googleapis.com/reusedata/gencode_annotation/genc...
#>   BFC4552 https://storage.googleapis.com/reusedata/gencode_annotation/genc...
#>   BFC4553 https://storage.googleapis.com/reusedata/gencode_transcriptome/g...
#>   ...     ...                                                                
#>   BFC4618 https://storage.googleapis.com/reusedata/STAR_index/GRCh38.GENCO...
#>   BFC4619 https://storage.googleapis.com/reusedata/STAR_index/GRCh38.GENCO...
#>   BFC4620 https://storage.googleapis.com/reusedata/STAR_index/GRCh38.GENCO...
#>   BFC4621 https://storage.googleapis.com/reusedata/STAR_index/GRCh38.GENCO...
#>   BFC4622 https://storage.googleapis.com/reusedata/STAR_index/GRCh38.GENCO...
dataSearch("#gatk")
#> dataHub with 0 records
#> cache path:  /home/qhu/.cache/R/ReUseData 
#> # dataUpdate() to update the local data cache
#> # dataSearch() to query a specific dataset
#> # Additional information can be retrieved using: 
#> # dataNames(), dataParams(), dataNotes(), dataPaths(), dataTag() or mcols()